Scope JSON vulnerabilities to the scanned repository version
Filter OSV reports by repository version so another index cannot show
vulns until it is scanned. Move the Warehouse trim helper out of
Django-backed utils so unit tests collect without loading apps.

Assisted By: Cursor Grok 4.6
